@@ -1,35 +1,55 @@
-# -*- mode: ruby; encoding: UTF-8 -*-
-source 'http://rubygems.org'
+# -*- ruby -*-
+source 'https://rubygems.org'
 
-gem 'rails', '3.2.2'
-gem 'sqlite3', group: [:test, :development]
-gem 'pg', group: :production
-gem 'paperclip'
-gem 'acts-as-taggable-on'
-gem 'kaminari'
 
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails', "  ~> 3.2.3"
-  gem 'coffee-rails', "~> 3.2.1"
-  gem 'uglifier', '>=1.0.3'
-  gem 'therubyracer'
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
+gem 'rails', '4.2.0'
+
+gem 'pg'
+
+gem 'jbuilder', '~> 2.0'
+
+# bundle exec rake doc:rails generates the API under doc/api.
+gem 'sdoc', '~> 0.4.0', group: :doc
+
+# Use ActiveModel has_secure_password
+# gem 'bcrypt', '~> 3.1.7'
+
+# Use Unicorn as the app server
+# gem 'unicorn'
+
+# Use Capistrano for deployment
+# gem 'capistrano-rails', group: :development
+
+gem 'carrierwave', github: 'carrierwaveuploader/carrierwave'
+gem 'mini_magick', '~> 4.0'
+gem 'kaminari', '~> 0.16.0'
+gem 'rails-i18n'
+gem 'simple_form'
+#gem 'formtastic'
+
+group :development, :test do
+  gem 'byebug'
+  gem 'pry-rails'
+  gem 'pry-git'
+  gem 'quiet_assets'
+  gem 'spring'
+  gem 'web-console', '~> 2.0'
 end
 
-gem 'jquery-rails'
+group :development, :test do
+  gem 'rspec-rails', '~> 3.0'
+  gem 'factory_girl_rails', '~> 4.5'
+end
 
-# Use unicorn as the web server
-#gem 'unicorn'
+# assets
 
-# Deploy with Capistrano
-gem 'capistrano'
-gem 'capistrano-unicorn'
+gem 'sass-rails', '~> 5.0'
+gem 'uglifier', '>= 1.3.0'
+gem 'coffee-rails', '~> 4.1.0'
+# gem 'therubyracer', platforms: :ruby
+# gem 'jquery-rails'
 
-# To use debugger
-# gem 'ruby-debug19', :require => 'ruby-debug'
-
-gem 'rspec-rails'
-group :test do
-  gem 'factory_girl_rails'
+source 'https://rails-assets.org' do
+  gem 'rails-assets-normalize-scss';
 end
