@@ -9,55 +9,19 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20111106180155) do
+ActiveRecord::Schema.define(version: 20150124080911) do
 
-  create_table "complaints", :force => true do |t|
-    t.integer  "entry_id",   :null => false
-    t.string   "ip_address", :null => false
-    t.datetime "created_at"
-    t.datetime "updated_at"
-  end
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
 
-  add_index "complaints", ["entry_id"], :name => "index_complaints_on_entry_id"
-  add_index "complaints", ["ip_address"], :name => "index_complaints_on_ip_address"
-
-  create_table "entries", :force => true do |t|
-    t.text     "body",               :default => "",           :null => false
-    t.string   "photo_file_name",    :default => "",           :null => false
-    t.string   "photo_content_type", :default => "",           :null => false
-    t.integer  "photo_file_size",    :default => 0,            :null => false
-    t.datetime "photo_updated_at"
-    t.string   "photo_fingerprint"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-    t.integer  "parent_id"
-    t.integer  "replies_count",      :default => 0,            :null => false
-    t.integer  "complaints_count",   :default => 0,            :null => false
-    t.boolean  "censored",           :default => false,        :null => false
-    t.string   "storage",            :default => "filesystem", :null => false
-  end
-
-  add_index "entries", ["created_at"], :name => "index_entries_on_created_at"
-  add_index "entries", ["parent_id"], :name => "index_entries_on_parent_id"
-  add_index "entries", ["photo_fingerprint"], :name => "index_entries_on_photo_fingerprint"
-
-  create_table "taggings", :force => true do |t|
-    t.integer  "tag_id"
-    t.integer  "taggable_id"
-    t.string   "taggable_type"
-    t.integer  "tagger_id"
-    t.string   "tagger_type"
-    t.string   "context"
-    t.datetime "created_at"
-  end
-
-  add_index "taggings", ["tag_id"], :name => "index_taggings_on_tag_id"
-  add_index "taggings", ["taggable_id", "taggable_type", "context"], :name => "index_taggings_on_taggable_id_and_taggable_type_and_context"
-
-  create_table "tags", :force => true do |t|
-    t.string "name"
+  create_table "entries", force: :cascade do |t|
+    t.string   "title"
+    t.string   "body"
+    t.json     "images"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
 end
