@@ -1,29 +1,15 @@
-// This is a manifest file that'll be compiled into including all the files listed below.
-// Add new JavaScript/Coffee code in separate files in this directory and they'll automatically
-// be included in the compiled file accessible from http://example.com/assets/application.js
-// It's not advisable to add code directly here, but if you do, it'll appear at the bottom of the
-// the compiled file.
+// This is a manifest file that'll be compiled into application.js, which will include all the files
+// listed below.
 //
-//= require jquery
-//= require jquery_ujs
-//= require entries
-//= require tags
-//= require jquery.lazyload2
-//= require ga
-//= require_self
-if (!window.console) console = {log: function(){}};
-jQuery(function($) {
-  $(document).lazyload2('.lazyload');
-  function autopagerHandler(e) {
-    $(e.target).lazyload2('.lazyload');
-  }
-  if (document.body && document.body.addEventListener) {
-    $([
-      'AutoPagerize_DOMNodeInserted',
-      'AutoPagerAfterInsert',
-      'AutoPatchWork.DOMNodeInserted'
-    ]).each(function() {
-      document.body.addEventListener(this, autopagerHandler, false);
-    });
-  }
-});
+// Any JavaScript/Coffee file within this directory, lib/assets/javascripts, vendor/assets/javascripts,
+// or any plugin's vendor/assets/javascripts directory can be referenced here using a relative path.
+//
+// It's not advisable to add code directly here, but if you do, it'll appear at the bottom of the
+// compiled file.
+//
+// Read Sprockets README (https://github.com/sstephenson/sprockets#sprockets-directives) for details
+// about supported directives.
+//
+// require jquery
+// require jquery_ujs
+//= require_tree .
